requirements.txt
Django
djangorestframework
psycopg2
BeautifulSoup
requests
lxml
#django-heroku
#gunicorn